+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - logger: Add time prefix (martadinata666)
 
+### Fixed
+
+- database: Fix video duration
+
 ## [1.10.0] - 2023-06-13
 
 ### Added